#19 Naming conventions
