+++ reader/src/main/scala/za/co/absa/kafkacase/reader/Reader.scala
@@ -14,14 +14,6 @@
  * limitations under the License.
  */
 
-package za.co.absa.KafkaCase.Writer
+package za.co.absa.kafkacase.reader
 
-trait Writer[TType] extends AutoCloseable {
-  def Write(key: String, value: TType): Unit
-  def Flush(): Unit
-
-  def WriteSync(key: String, value: TType): Unit = {
-    Write(key, value)
-    Flush()
-  }
-}
+trait Reader[TType] extends Iterator[(String, Either[String, TType])] with AutoCloseable
