+++ reader/src/main/scala/za/co/absa/kafkacase/reader/ReaderImpl.scala
@@ -14,13 +14,13 @@
  * limitations under the License.
  */
 
-package za.co.absa.KafkaCase.Reader
+package za.co.absa.kafkacase.reader
 
 import io.circe.Decoder
 import io.circe.jawn.decode
 import org.slf4j.LoggerFactory
 import org.apache.kafka.clients.consumer.{ConsumerRecord, KafkaConsumer}
-import za.co.absa.KafkaCase.Reader.ReaderImpl.{DEFAULT_TIMEOUT, log}
+import za.co.absa.kafkacase.reader.ReaderImpl.{DEFAULT_TIMEOUT, log}
 
 import java.time.Duration
 import java.util
